♻️ refactor: 스크롤 로직 분리 및 좌석 페이지에 스크롤 기능 추가
