+++ src/components/pages/seats/SeatGrid.tsx
@@ -1,3 +1,8 @@
+import ScrollProvider from "@/components/Layout/ScrollProvider";
+import useTabDrag from "@/components/commons/Tab/useTabDrag";
+import { containerRefAtom } from "@/store/scrollAtom";
+import { useAtomValue } from "jotai";
+
 import SeatButton from "./SeatButton";
 import useGetAllSeatsData from "./hooks/useGetAllSeatsData";
 import type { SeatData } from "./types";
@@ -5,21 +10,35 @@ import type { SeatData } from "./types";
 export default function SeatGrid() {
   const allSeatsData = useGetAllSeatsData();
 
+  const containerRef = useAtomValue(containerRefAtom);
+  const { handleMouseDown, handleMouseMove, handleMouseUpOrLeave } =
+    useTabDrag(containerRef);
+
   return (
-    <div className="mb-200 overflow-x-auto md:mx-60">
-      <div className="mx-16 my-28 grid w-[668px] auto-rows-auto grid-cols-2 gap-40 md:my-70 md:w-[1004px] xl:flex-shrink-0">
-        <SeatBlock seatsData={allSeatsData.A} />
-        <SeatBlock seatsData={allSeatsData.B} />
-        <SeatBlock seatsData={allSeatsData.C} />
-        <SeatBlock seatsData={allSeatsData.D} />
-        <SeatBlock seatsData={allSeatsData.E} />
-        <SeatBlock seatsData={allSeatsData.F} />
-        <SeatBlock seatsData={allSeatsData.G} />
-        <SeatBlock seatsData={allSeatsData.H} />
-        <SeatBlock seatsData={allSeatsData.I} />
-        <SeatBlock seatsData={allSeatsData.J} />
+    <ScrollProvider>
+      {/* eslint-disable-next-line */}
+      <div
+        className="mb-200 overflow-x-auto md:mx-60"
+        ref={containerRef}
+        onMouseDown={handleMouseDown}
+        onMouseMove={handleMouseMove}
+        onMouseLeave={handleMouseUpOrLeave}
+        onMouseUp={handleMouseUpOrLeave}
+      >
+        <div className="mx-16 my-28 grid w-[668px] auto-rows-auto grid-cols-2 gap-40 md:my-70 md:w-[1004px] xl:flex-shrink-0">
+          <SeatBlock seatsData={allSeatsData.A} />
+          <SeatBlock seatsData={allSeatsData.B} />
+          <SeatBlock seatsData={allSeatsData.C} />
+          <SeatBlock seatsData={allSeatsData.D} />
+          <SeatBlock seatsData={allSeatsData.E} />
+          <SeatBlock seatsData={allSeatsData.F} />
+          <SeatBlock seatsData={allSeatsData.G} />
+          <SeatBlock seatsData={allSeatsData.H} />
+          <SeatBlock seatsData={allSeatsData.I} />
+          <SeatBlock seatsData={allSeatsData.J} />
+        </div>
       </div>
-    </div>
+    </ScrollProvider>
   );
 }
